+++ message-config/src/main/java/message/config/auto/AutoConfigurationCreator.java
@@ -1,6 +1,15 @@
 package message.config.auto;
 
-import message.base.utils.GenericsUtils;
+import java.beans.PropertyDescriptor;
+import java.io.Serializable;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Properties;
+import java.util.Set;
+import message.base.utils.GenericUtils;
 import message.base.utils.PropertyPlaceholderHelper;
 import message.base.utils.StringUtils;
 import message.config.SystemConfig;
@@ -8,100 +17,98 @@ import message.config.auto.annotations.ConfigurationProperties;
 import message.config.loader.properties.Configuration;
 import message.config.loader.properties.ConfigurationLoader;
 import message.config.loader.properties.PropertiesConfiguration;
+import org.apache.commons.lang.ArrayUtils;
 import org.springframework.beans.BeanUtils;
-import org.springframework.core.env.Environment;
-import org.springframework.core.env.StandardEnvironment;
+import org.springframework.beans.BeanWrapper;
+import org.springframework.beans.BeanWrapperImpl;
 import org.springframework.core.io.DefaultResourceLoader;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.core.io.support.PropertiesLoaderUtils;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.Serializable;
-import java.lang.reflect.Field;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Properties;
-import java.util.Set;
-
 /**
  * 自动装配生成器.
  *
  * @author sunhao(sunhao.java@gmail.com)
- * @version V1.0, 16/1/21 下午4:37
+ * @version V1.0, 2016-1-26 14:17
  */
 public class AutoConfigurationCreator {
     private static final List<? extends Class<? extends Serializable>> SIMPLE_CLASS = Arrays.asList(Boolean.class, boolean.class, Long.class, long.class,
             Integer.class, int.class, String.class, Double.class, double.class);
     private ResourceLoader resourceLoader = new DefaultResourceLoader();
-    private Environment environment = new StandardEnvironment();
-    private Configuration _configuration_;
 
-    public AutoConfigurationCreator(Configuration configuration) {
-        if (configuration == null)
-            _configuration_ = new PropertiesConfiguration(ConfigurationLoader.getConfigProperties());
-        else
-            _configuration_ = configuration;
+    private AutoConfigurationCreator() {
     }
 
-    public <T> T createBean(Class<T> clazz, ConfigurationProperties configuration) throws Exception {
-        String prefix = configuration.prefix();
-        Configuration _configuration = findConfiguration(configuration);
-        return (T) generateBean(prefix, clazz, _configuration);
+    private <T> T generateConfigurationBean(Class<T> clazz, ConfigurationProperties annotation) throws Exception {
+        Configuration configuration = loadProp(annotation.locations());
+        return generateObject(annotation.prefix(), clazz, configuration);
     }
 
-    private Object generateBean(String prefix, Class<?> clazz, Configuration _configuration) throws IllegalAccessException {
+    private <T> T generateObject(String prefix, Class<T> clazz, Configuration configuration) throws Exception {
         Object object = BeanUtils.instantiate(clazz);
+        BeanWrapper beanWrapper = new BeanWrapperImpl(object);
 
-        Field[] fields = clazz.getDeclaredFields();
+        PropertyDescriptor[] descriptors = beanWrapper.getPropertyDescriptors();
+        for (PropertyDescriptor descriptor : descriptors) {
+            if (descriptor.getWriteMethod() == null) {
+                continue;
+            }
 
-        for (Field field : fields) {
-            field.setAccessible(true);
+            String name = descriptor.getName();
+            Class<?> type = descriptor.getPropertyType();
+            Method method = descriptor.getReadMethod();
 
-            String fieldName = field.getName();
-            Class<?> type = field.getType();
-
-            String key = prefix + "." + fieldName;
-
+            String key = prefix + "." + name;
+            Object value;
             if (isSimpleType(type)) {
-                Object value = generateForSimpleType(key, type, _configuration);
-                if (value != null) {
-                    field.set(object, value);
-                }
+                value = getValueForSimpleType(key, type, configuration);
             } else if (Map.class.equals(type)) {
-                Map<String, Object> map = generateForMap(key, field, _configuration);
-                if (map != null && !map.isEmpty()) {
-                    field.set(object, map);
-                }
+                value = getValueForMap(key, method, configuration);
             } else {
-                // 对象
-                Object object_ = generateBean(key, type, _configuration);
-                field.set(object, object_);
+                value = generateObject(key, type, configuration);
+            }
+
+            if (value != null) {
+                beanWrapper.setPropertyValue(name, value);
             }
         }
 
-        return object;
+        return (T) object;
+    }
+
+    private Configuration loadProp(String... configLocations) throws Exception {
+        if (ArrayUtils.isEmpty(configLocations)) {
+            return new PropertiesConfiguration(ConfigurationLoader.getConfigProperties());
+        }
+
+        Properties prop = new Properties();
+        for (String location : configLocations) {
+            location = PropertyPlaceholderHelper.replacePlaceholder(location, true, SystemConfig.getAllConfigs());
+
+            Resource resource = this.resourceLoader.getResource(location);
+            PropertiesLoaderUtils.fillProperties(prop, resource);
+        }
+
+        return new PropertiesConfiguration(prop);
     }
 
     private boolean isSimpleType(Class<?> type) {
         return SIMPLE_CLASS.contains(type);
     }
 
-    private Object generateForSimpleType(String key, Class<?> type, Configuration _configuration) {
+    private Object getValueForSimpleType(String key, Class<?> type, Configuration configuration) {
         try {
             if (Boolean.class.equals(type) || boolean.class.equals(type)) {
-                return _configuration.getBoolean(key);
+                return configuration.getBoolean(key);
             } else if (Long.class.equals(type) || long.class.equals(type)) {
-                return _configuration.getLong(key);
+                return configuration.getLong(key);
             } else if (Integer.class.equals(type) || int.class.equals(type)) {
-                return _configuration.getInt(key);
+                return configuration.getInt(key);
             } else if (String.class.equals(type)) {
-                return _configuration.getString(key);
+                return configuration.getString(key);
             } else if (Double.class.equals(type) || double.class.equals(type)) {
-                return _configuration.getDouble(key);
+                return configuration.getDouble(key);
             }
         } catch (Exception e) {
             return null;
@@ -110,14 +117,14 @@ public class AutoConfigurationCreator {
         return null;
     }
 
-    private Map<String, Object> generateForMap(String prefix, Field field, Configuration _configuration) {
-        if (!Map.class.equals(field.getType()) || !String.class.equals(GenericsUtils.getFieldGenericType0(field))) {
+    private Map<String, Object> getValueForMap(String prefix, Method method, Configuration configuration) throws Exception {
+        if (!Map.class.equals(method.getReturnType()) || !String.class.equals(GenericUtils.getGenericParameter0(method))) {
             return null;
         }
 
         Map<String, Object> map = new HashMap<>();
-        Class<?> secondGenericClazz = GenericsUtils.getFieldGenericType(field, 1);
-        Set<String> keys = _configuration.subset(prefix).getKeys();
+        Class<?> secondGenericClazz = GenericUtils.getGenericParameter(method, 1);
+        Set<String> keys = configuration.subset(prefix).getKeys();
         for (String key : keys) {
             String[] temp = StringUtils.split(key, ".");
             if (temp.length != 2) {
@@ -125,7 +132,7 @@ public class AutoConfigurationCreator {
             }
 
             String keyInMap = temp[0];
-            Object object = generateObject(prefix + "." + keyInMap, secondGenericClazz, _configuration);
+            Object object = generateObject(prefix + "." + keyInMap, secondGenericClazz, configuration);
             if (object != null)
                 map.put(keyInMap, object);
         }
@@ -133,60 +140,27 @@ public class AutoConfigurationCreator {
         return map;
     }
 
-    private Object generateObject(String prefix, Class<?> clazz, Configuration _configuration) {
-        Field[] fields = clazz.getDeclaredFields();
+    public static class Builder<T> {
+        private AutoConfigurationCreator creator = new AutoConfigurationCreator();
+        private Class<T> clazz;
+        private ConfigurationProperties annotation;
 
-        Object object = BeanUtils.instantiate(clazz);
-
-        try {
-            for (Field field : fields) {
-                field.setAccessible(true);
-
-                String fieldName = field.getName();
-                Class<?> type = field.getType();
-
-                String key = prefix + "." + fieldName;
-
-                if (isSimpleType(type)) {
-                    Object value = generateForSimpleType(key, type, _configuration);
-                    if (value != null) {
-                        field.set(object, value);
-                    }
-                } else if (Map.class.equals(type)) {
-                    Map<String, Object> map = generateForMap(key, field, _configuration);
-                    if (map != null && !map.isEmpty())
-                        field.set(object, map);
-                } else {
-                    // 对象
-                    Object object_ = generateObject(key, type, _configuration);
-                    field.set(object, object_);
-                }
-            }
-        } catch (Exception e) {
-            return null;
+        public Builder<T> setClazz(Class<T> clazz) {
+            this.clazz = clazz;
+            return this;
         }
 
-        return object;
-    }
-
-    private Configuration findConfiguration(ConfigurationProperties properties) throws FileNotFoundException {
-        String[] locations = properties.locations();
-        if (locations == null || locations.length == 0) {
-            return _configuration_;
+        public Builder<T> setAnnotation(ConfigurationProperties annotation) {
+            this.annotation = annotation;
+            return this;
         }
 
-        Properties prop = new Properties();
-        for (String location : locations) {
-            location = PropertyPlaceholderHelper.replacePlaceholder(location, location, SystemConfig.getAllConfigs());
-
-            Resource resource = this.resourceLoader.getResource(this.environment.resolvePlaceholders(location));
+        public T build() {
             try {
-                PropertiesLoaderUtils.fillProperties(prop, resource);
-            } catch (IOException e) {
-                throw new FileNotFoundException("File '" + location + "' does not exist");
+                return creator.generateConfigurationBean(clazz, annotation);
+            } catch (Exception e) {
+                return null;
             }
         }
-
-        return new PropertiesConfiguration(prop);
     }
 }