+++ message-mongodb/src/main/java/message/mongodb/annotations/EnableMongo.java
@@ -33,5 +33,13 @@ public @interface EnableMongo {
      * @return
      * @see org.springframework.stereotype.Repository
      */
-    String basePackage();
+    String[] basePackage() default {};
+
+    /**
+     * 含有{@link org.springframework.data.mongodb.core.mapping.Document }注解的dao类所在的包路径
+     *
+     * @return
+     * @see org.springframework.data.mongodb.core.mapping.Document
+     */
+    String[] domainPackage() default {};
 }