+++ message-mongodb/src/main/java/message/mongodb/annotations/MongoBeanDefinitionRegistrar.java
@@ -1,7 +1,9 @@
 package message.mongodb.annotations;
 
-import message.mongodb.core.MongoDataSourceBeanDefinitionFactory;
 import message.base.template.ThymeleafTemplateResource;
+import message.base.utils.StringUtils;
+import message.mongodb.core.MongoDataSourceBeanDefinitionFactory;
+import org.apache.commons.lang.ArrayUtils;
 import org.springframework.beans.factory.support.BeanDefinitionReader;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.beans.factory.xml.XmlBeanDefinitionReader;
@@ -9,8 +11,11 @@ import org.springframework.context.annotation.ImportBeanDefinitionRegistrar;
 import org.springframework.core.annotation.AnnotationAttributes;
 import org.springframework.core.io.Resource;
 import org.springframework.core.type.AnnotationMetadata;
+import org.springframework.core.type.StandardAnnotationMetadata;
 import org.springframework.util.Assert;
+import org.springframework.util.ClassUtils;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -21,31 +26,39 @@ import java.util.Map;
  * @version V1.0, 16/1/21 下午10:15
  */
 public class MongoBeanDefinitionRegistrar implements ImportBeanDefinitionRegistrar {
-    public static final String DATASOURCE_NAME_ATTRIBUTE_NAME = "name";
     private static final String DATA_SOURCE_ATTRIBUTE_NAME = "dataSource";
     private static final String BASE_PACKAGE_ATTRIBUTE_NAME = "basePackage";
+    private static final String DOMAIN_PACKAGE_ATTRIBUTE_NAME = "domainPackage";
 
     @Override
     public void registerBeanDefinitions(AnnotationMetadata importingClassMetadata, BeanDefinitionRegistry registry) {
         AnnotationAttributes attributes = AnnotationAttributes.fromMap(importingClassMetadata.getAnnotationAttributes(EnableMongo.class.getName(), false));
         Assert.notNull(attributes, String.format("@%s is not present on importing class '%s' as expected", EnableMongo.class.getName(), importingClassMetadata.getClassName()));
 
-        String dataSource = attributes.getString(DATA_SOURCE_ATTRIBUTE_NAME);
-
         // 注册数据源
-        String name = attributes.getString(DATASOURCE_NAME_ATTRIBUTE_NAME);
-        registry.registerBeanDefinition(name, new MongoDataSourceBeanDefinitionFactory(name).build());
+        String dataSource = attributes.getString(DATA_SOURCE_ATTRIBUTE_NAME);
+        registry.registerBeanDefinition(dataSource, new MongoDataSourceBeanDefinitionFactory(dataSource).build());
 
-        String basePackage = attributes.getString(BASE_PACKAGE_ATTRIBUTE_NAME);
+        String[] basePackage = attributes.getStringArray(BASE_PACKAGE_ATTRIBUTE_NAME);
+        String[] domainPackage = attributes.getStringArray(DOMAIN_PACKAGE_ATTRIBUTE_NAME);
+
+        Class<?> introspectedClass = ((StandardAnnotationMetadata) importingClassMetadata).getIntrospectedClass();
+        if (ArrayUtils.isEmpty(basePackage) && introspectedClass != null) {
+            basePackage = Arrays.asList(ClassUtils.getPackageName(introspectedClass)).toArray(new String[1]);
+        }
+        if (ArrayUtils.isEmpty(domainPackage) && introspectedClass != null) {
+            domainPackage = Arrays.asList(ClassUtils.getPackageName(introspectedClass)).toArray(new String[1]);
+        }
 
         BeanDefinitionReader beanDefinitionReader = new XmlBeanDefinitionReader(registry);
-        beanDefinitionReader.loadBeanDefinitions(loadBeanDefinitions(dataSource, basePackage));
+        beanDefinitionReader.loadBeanDefinitions(loadBeanDefinitions(dataSource, basePackage, domainPackage));
     }
 
-    private Resource loadBeanDefinitions(String dataSource, String basePackage) {
+    private Resource loadBeanDefinitions(String dataSource, String[] basePackage, String[] domainPackage) {
         Map<String, Object> context = new HashMap<>();
         context.put("dataSource", dataSource);
-        context.put("basePackage", basePackage);
+        context.put("basePackage", StringUtils.join(Arrays.asList(basePackage), ","));
+        context.put("domainPackage", StringUtils.join(Arrays.asList(domainPackage), ","));
 
         return new ThymeleafTemplateResource("META-INF/template/mongo.xml", context, "xml");
     }