+++ message-mvc/src/main/java/message/mvc/commons/WebInput.java
@@ -1,12 +1,12 @@
 package message.mvc.commons;
 
 import message.base.bean.Constants;
-import message.mvc.editor.CustomDateEditor;
 import message.base.utils.RequestUtils;
 import message.base.utils.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
+import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.web.bind.ServletRequestDataBinder;
 
 import javax.servlet.http.Cookie;
@@ -21,10 +21,13 @@ import java.util.regex.Pattern;
 
 /**
  * 封装HttpServletRequest
+ *
  * @author sunhao(sunhao.java@gmail.com)
  */
 public class WebInput {
-    /** Default command name used for binding command objects: "command" */
+    /**
+     * Default command name used for binding command objects: "command"
+     */
     public static final String DEFAULT_COMMAND_NAME = "command";
 
     private static final Logger log = LoggerFactory.getLogger(WebInput.class);
@@ -38,7 +41,7 @@ public class WebInput {
     private static final Integer DEFAULT_SESSION_LIFE = 20 * 60 * 1000;
     private HttpServletRequest request;
 
-    public WebInput(HttpServletRequest request){
+    public WebInput(HttpServletRequest request) {
         this.request = request;
     }
 
@@ -50,7 +53,7 @@ public class WebInput {
         return StringUtils.isEmpty(this.request.getParameter(name)) ? defaultValue : this.request.getParameter(name);
     }
 
-    public String getString(String name){
+    public String getString(String name) {
         return StringUtils.trimToNull(this.request.getParameter(name));
     }
 
@@ -66,7 +69,7 @@ public class WebInput {
         return result;
     }
 
-    public Integer getInt(String name, Integer defaultValue){
+    public Integer getInt(String name, Integer defaultValue) {
         Integer result = defaultValue;
         try {
             result = this.request.getParameter(name) != null ?
@@ -78,9 +81,9 @@ public class WebInput {
         return result;
     }
 
-    public Integer getInt(String name){
+    public Integer getInt(String name) {
         try {
-            if(StringUtils.isEmpty(this.request.getParameter(name))){
+            if (StringUtils.isEmpty(this.request.getParameter(name))) {
                 return null;
             } else {
                 return Integer.parseInt(this.request.getParameter(name));
@@ -90,7 +93,7 @@ public class WebInput {
         }
     }
 
-    public long getLong(String name, long defaultValue){
+    public long getLong(String name, long defaultValue) {
         long result = defaultValue;
         try {
             result = Long.parseLong(this.request.getParameter(name));
@@ -101,7 +104,7 @@ public class WebInput {
         return result;
     }
 
-    public Long getLong(String name, Long defaultValue){
+    public Long getLong(String name, Long defaultValue) {
         Long result = defaultValue;
         try {
             result = Long.valueOf(this.request.getParameter(name));
@@ -112,12 +115,12 @@ public class WebInput {
         return result;
     }
 
-    public Long getLong(String name){
+    public Long getLong(String name) {
         String paramStr = this.request.getParameter(name);
         return StringUtils.isEmpty(paramStr) ? null : Long.valueOf(paramStr);
     }
 
-    public double getDouble(String name, double defaultValue){
+    public double getDouble(String name, double defaultValue) {
         double result = defaultValue;
         try {
             result = Double.parseDouble(this.request.getParameter(name));
@@ -131,14 +134,14 @@ public class WebInput {
     public boolean getBoolean(String name, boolean defaultValue) {
         boolean result = defaultValue;
         String s = StringUtils.trimToNull(this.request.getParameter(name));
-        if(s == null){
+        if (s == null) {
             return result;
         }
         s = s.toLowerCase();
 
-        if("yes".equals(s) || "true".equals(s) || "1".equals(s) || "t".equals(s)) {
+        if ("yes".equals(s) || "true".equals(s) || "1".equals(s) || "t".equals(s)) {
             result = true;
-        } else if("no".equals(s) || "false".equals(s) || "0".equals(s) || "f".equals(s)) {
+        } else if ("no".equals(s) || "false".equals(s) || "0".equals(s) || "f".equals(s)) {
             result = false;
         }
 
@@ -150,7 +153,7 @@ public class WebInput {
         Date result = defaultValue;
         try {
             String value = this.request.getParameter(name);
-            if(StringUtils.isNotBlank(value)) {
+            if (StringUtils.isNotBlank(value)) {
                 result = format.parse(value);
             }
         } catch (ParseException e) {
@@ -172,17 +175,17 @@ public class WebInput {
         }
     }
 
-    public String[] getStrings(String name){
+    public String[] getStrings(String name) {
         return this.request.getParameterValues(name);
     }
 
-    public int[] getInts(String name, int defaultValue){
+    public int[] getInts(String name, int defaultValue) {
         String[] values = this.getStrings(name);
-        if(values == null || values.length < 1){
+        if (values == null || values.length < 1) {
             return null;
         }
         int[] results = new int[values.length];
-        for(int i = 0; i < values.length; i++) {
+        for (int i = 0; i < values.length; i++) {
             results[i] = defaultValue;
             try {
                 results[i] = Integer.parseInt(values[i]);
@@ -194,17 +197,17 @@ public class WebInput {
         return results;
     }
 
-    public int[] getInts(String name){
+    public int[] getInts(String name) {
         return this.getInts(name, 0);
     }
 
-    public long[] getLongs(String name, long defaultValue){
+    public long[] getLongs(String name, long defaultValue) {
         String[] values = this.getStrings(name);
-        if(values == null || values.length < 1){
+        if (values == null || values.length < 1) {
             return null;
         }
         long[] results = new long[values.length];
-        for(int i = 0; i < values.length; i++) {
+        for (int i = 0; i < values.length; i++) {
             results[i] = defaultValue;
             try {
                 results[i] = Long.parseLong(values[i]);
@@ -216,39 +219,39 @@ public class WebInput {
         return results;
     }
 
-    public long[] getLongs(String name){
+    public long[] getLongs(String name) {
         return this.getLongs(name, 0L);
     }
 
-    public Long[] getLongObjects(String name, Long defaultValue){
+    public Long[] getLongObjects(String name, Long defaultValue) {
         long[] values = this.getLongs(name, defaultValue);
-        if(values == null || values.length < 1){
+        if (values == null || values.length < 1) {
             return null;
         }
         Long[] results = new Long[values.length];
-        for(int i = 0; i < values.length; i++){
+        for (int i = 0; i < values.length; i++) {
             results[i] = new Long(values[i]);
         }
 
         return results;
     }
 
-    public Long[] getLongObjects(String name){
+    public Long[] getLongObjects(String name) {
         return this.getLongObjects(name, 0L);
     }
 
-    public String getCookieValue(String name){
+    public String getCookieValue(String name) {
         Cookie[] cs = this.request.getCookies();
 
-        if(cs == null || cs.length < 1){
+        if (cs == null || cs.length < 1) {
             return null;
         }
 
-        for(int i = 0; i < cs.length; i++){
+        for (int i = 0; i < cs.length; i++) {
             Cookie c = cs[i];
             String key = c.getName();
             String value = c.getValue();
-            if(name.equals(key)){
+            if (name.equals(key)) {
                 return value;
             }
         }
@@ -256,34 +259,34 @@ public class WebInput {
         return null;
     }
 
-    public HttpSession getSession(){
+    public HttpSession getSession() {
         return this.request.getSession();
     }
 
-    public void setMaxInactiveInterval(HttpSession session, Integer maxTime){
+    public void setMaxInactiveInterval(HttpSession session, Integer maxTime) {
         session.setMaxInactiveInterval(maxTime == null ? DEFAULT_SESSION_LIFE : maxTime);
     }
 
-    public String getClientIP(){
+    public String getClientIP() {
         return this.request.getRemoteAddr();
     }
 
-    public void setAttribute(String name, Object value){
+    public void setAttribute(String name, Object value) {
         request.setAttribute(name, value);
     }
 
-    public Object getAttribute(String name){
+    public Object getAttribute(String name) {
         return this.request.getAttribute(name);
     }
 
     /**
      * 从request中获取一个bean
      *
-     * @param beanClass     bean的class
+     * @param beanClass bean的class
      * @param <T>
      * @return
      */
-    public <T> T getBean(Class<T> beanClass){
+    public <T> T getBean(Class<T> beanClass) {
         Object command = BeanUtils.instantiate(beanClass);
         try {
             bind(request, command);
@@ -296,6 +299,7 @@ public class WebInput {
 
     /**
      * Bind request parameters onto the given command bean
+     *
      * @param request request from which parameters will be bound
      * @param command command object, that must be a JavaBean
      * @throws Exception in case of invalid state or arguments
@@ -314,6 +318,7 @@ public class WebInput {
      * <p>The default implementation creates a standard ServletRequestDataBinder,
      * and invokes <code>initBinder</code>. Note that <code>initBinder</code>
      * will not be invoked if you override this method!
+     *
      * @param request current HTTP request
      * @param command the command to bind onto
      * @return the new binder instance
@@ -332,11 +337,12 @@ public class WebInput {
 
     /**
      * get date format pattern for what parameter in request
+     *
      * @param urlParams
      * @return
      */
-    private String getDateFormatPattern(List<String> urlParams){
-        if(urlParams.isEmpty()){
+    private String getDateFormatPattern(List<String> urlParams) {
+        if (urlParams.isEmpty()) {
             log.debug("this url paramter is null!");
             return Constants.SIMPLE_DATE_FORMAT;
         }
@@ -346,14 +352,14 @@ public class WebInput {
         Pattern p2 = Pattern.compile("(\\d){2,4}[-](\\d){1,2}[-](\\d){1,2}");
         //hh:mm yyyy-MM-dd
         Pattern p3 = Pattern.compile("(\\d){1,2}[:](\\d){1,2} (\\d){2,4}[-](\\d){1,2}[-](\\d){1,2}");
-        for(String up : urlParams){
-            if(p1.matcher(up).matches()){
+        for (String up : urlParams) {
+            if (p1.matcher(up).matches()) {
                 return Constants.SIMPLE_DATE_FORMAT;
             }
-            if(p2.matcher(up).matches()){
+            if (p2.matcher(up).matches()) {
                 return Constants.DATE_FORMAT;
             }
-            if(p3.matcher(up).matches()){
+            if (p3.matcher(up).matches()) {
                 return Constants.DATE_FORMAT_;
             }
         }