+++ message-mybatis/src/main/java/message/mybatis/configs/MyBatisBeanDefinitionRegistrar.java
@@ -1,5 +1,6 @@
 package message.mybatis.configs;
 
+import java.util.Arrays;
 import message.base.utils.StringUtils;
 import message.mybatis.configs.annotations.EnableMyBatis;
 import message.mybatis.helper.MySQLSqlHelper;
@@ -7,6 +8,7 @@ import message.mybatis.helper.OracleSqlHelper;
 import message.mybatis.key.generic.MySQLMaxValueIncrementer;
 import message.mybatis.key.sequence.OracleSequenceMaxValueIncrementer;
 import message.mybatis.type.TypeHandlerScanner;
+import org.apache.commons.lang.ArrayUtils;
 import org.flywaydb.core.Flyway;
 import org.mybatis.spring.SqlSessionFactoryBean;
 import org.mybatis.spring.mapper.MapperScannerConfigurer;
@@ -16,6 +18,7 @@ import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.context.annotation.ImportBeanDefinitionRegistrar;
 import org.springframework.core.annotation.AnnotationAttributes;
 import org.springframework.core.type.AnnotationMetadata;
+import org.springframework.core.type.StandardAnnotationMetadata;
 import org.springframework.jdbc.support.lob.DefaultLobHandler;
 import org.springframework.jdbc.support.lob.OracleLobHandler;
 import org.springframework.jdbc.support.nativejdbc.CommonsDbcpNativeJdbcExtractor;
@@ -57,6 +60,15 @@ public class MyBatisBeanDefinitionRegistrar implements ImportBeanDefinitionRegis
 
         String[] typeHandlersLocations = attributes.getStringArray(TYPE_HANDLERS_LOCATIONS_ATTRIBUTE_NAME);
         String[] basePackages = attributes.getStringArray(BASE_PACKAGES_ATTRIBUTE_NAME);
+
+        Class<?> introspectedClass = ((StandardAnnotationMetadata) importingClassMetadata).getIntrospectedClass();
+        if (ArrayUtils.isEmpty(basePackages) && introspectedClass != null) {
+            basePackages = Arrays.asList(ClassUtils.getPackageName(introspectedClass)).toArray(new String[1]);
+        }
+        if (ArrayUtils.isEmpty(typeHandlersLocations) && introspectedClass != null) {
+            typeHandlersLocations = Arrays.asList(ClassUtils.getPackageName(introspectedClass)).toArray(new String[1]);
+        }
+
         boolean useFlyway = attributes.getBoolean(USE_FLYWAY_ATTRIBUTE_NAME);
         String migration = attributes.getString(MIGRATION_ATTRIBUTE_NAME);