+++ message-redis/src/main/java/message/redis/core/connection/RedisDataSourceBeanDefinitionFactory.java
@@ -1,6 +1,5 @@
 package message.redis.core.connection;
 
-import message.config.SystemConfig;
 import message.config.auto.AutoConfigurationCreator;
 import message.config.auto.annotations.ConfigurationProperties;
 import message.redis.core.config.RedisProperties;
@@ -14,15 +13,17 @@ import org.springframework.beans.factory.support.BeanDefinitionBuilder;
  * @version V1.0, 16/1/21 下午4:10
  */
 public class RedisDataSourceBeanDefinitionFactory {
-    private static final String DATASOURCE_FILE_NAME = "redis.properties";
-
     private String dataSourceName;
     private RedisProperties redisProperties;
 
     public RedisDataSourceBeanDefinitionFactory(String dataSourceName) {
         this.dataSourceName = dataSourceName;
         try {
-            this.redisProperties = new AutoConfigurationCreator(SystemConfig.getFileConfiguration(DATASOURCE_FILE_NAME)).createBean(RedisProperties.class, RedisProperties.class.getAnnotation(ConfigurationProperties.class));
+            AutoConfigurationCreator.Builder<RedisProperties> builder = new AutoConfigurationCreator.Builder<>();
+            builder.setAnnotation(RedisProperties.class.getAnnotation(ConfigurationProperties.class));
+            builder.setClazz(RedisProperties.class);
+
+            this.redisProperties = builder.build();
         } catch (Exception e) {
             e.printStackTrace();
         }